+++ src/Module.EvcVerification/Global.cs
@@ -10,17 +10,17 @@ namespace Module.EvcVerification
         /// <summary>
         /// Defines the COR_ERROR_THRESHOLD
         /// </summary>
-        public const double COR_ERROR_THRESHOLD = 1.5;
+        public const decimal COR_ERROR_THRESHOLD = 1.5M;
 
         /// <summary>
         /// Defines the METER_DIS_ERROR_THRESHOLD
         /// </summary>
-        public const double METER_DIS_ERROR_THRESHOLD = 1;
+        public const decimal METER_DIS_ERROR_THRESHOLD = 1;
 
         /// <summary>
         /// Defines the PRESSURE_ERROR_TOLERANCE
         /// </summary>
-        public const double PRESSURE_ERROR_TOLERANCE = 1.0;
+        public const decimal PRESSURE_ERROR_TOLERANCE = 1.0M;
 
         /// <summary>
         /// Defines the PULSE_VARIANCE_THRESHOLD
@@ -30,17 +30,17 @@ namespace Module.EvcVerification
         /// <summary>
         /// Defines the SUPER_FACTOR_TOLERANCE
         /// </summary>
-        public const double SUPER_FACTOR_TOLERANCE = 0.3;
+        public const decimal SUPER_FACTOR_TOLERANCE = 0.3M;
 
         /// <summary>
         /// Defines the TEMP_ERROR_TOLERANCE
         /// </summary>
-        public const double TEMP_ERROR_TOLERANCE = 1.0;
+        public const decimal TEMP_ERROR_TOLERANCE = 1.0M;
 
         /// <summary>
         /// Defines the UNCOR_ERROR_THRESHOLD
         /// </summary>
-        public const double UNCOR_ERROR_THRESHOLD = 0.1;
+        public const decimal UNCOR_ERROR_THRESHOLD = 0.1M;
 
         #endregion
     }
